fix:  use arguments rather than options object
An options object suggest that some arguments might not be required.
In the case of this module all arguments are required and not passing any of them leads to very unexpected and dangerous results.

BREAKING CHANGE: Instead of an options object the function is now expecting the arguments to be passed individually.